Make % an allowed keyword character
I encountered a database column with a percentage sign in its name. When trying to list it using a
generic sql function that inserts a row with column names based on the keys given in a map, I found
that HugSQL would lex `:portion_%` into `:portion_` and then error out telling the user:

Parameter Mismatch: :portion_ parameter data not found.

This commit expands the characters allowed in keywords to include the percentage sign %.